Jacob - This fixes the images showing on all the items in the array.
I had to edit the picasso library a little. it is under the libs folder now. It requires all the dependecies in the lib folder to run and then the okhttp from the libs folder. Sorry, my file system got all messed up. but just put them into a library and set it as a dependecy for picasso to get it working.

It seems that the recycling is still a little weird, so i will work that out next.