retweets working correctly.
Jacob - Just so you know, there will be a lot of times right now when after you update, it is gonna crash. That is because i am editing the database and creating new columns and such that you don't have in the database yet. Just clear the data and it will let you log in again.