more stuff for second profile, added a default heaer for the drawer as well, just in case the user doesn't have one.
I think that the second profile stuff should be done and seems to be working for me... will need testing though of course. it will only background sync the current profile as well, i don't know how falcon does it. i doubt that it refreshes the timelines for the second profiles, but i bet it does mentions, so i will need to do that i think