some fixes.
What a convoluted way to make the popup work, but i was having problems with it again today. It had stopped clearing the top again, so it was popping up over the main activity! grrr buuutttt, i fixed it by calling this redirect intent which i was able to finish, then clear the top on the popup. should work perfectly every time now, don't know about any weird animations for some people, but they are dumb to change stuff like that, so i don't care.