reworked the live streaming again. the problem is that there is a possibility to miss tweets if you come into the app and have live streaming on but it isn't at the top of the list, so it doesn't automatically refresh. That means that you could possibly be behind on the tweets, but the live streaming won't stop and it will insert that tweet when it comes in, placing it at the end of the database and then i will read it as the last tweet and all of the ones before will be missed...
This should be gone now, because right after startup, i make calls to the api to refresh the timeline when the first new one from the streaming comes in. the user won't notice, but it should work to make sure all the tweets are put in correctly. the only thing is that that first one will take a few extra seconds to show up in the toast, but i don't care haha

just had to write out all my thoughts before i forgot them haha