lets finally stop delaying things on the timelines
It is plenty optimized now, delaying actions and creating/posting to handlers like that is hurting more than it helps. Especially since glide now manages the images instead of us doing that ourselves. The only thing that is a bit worse is that it will eagerly load quoted tweets a bit sooner, but this should be a change for the better, not worse